pyproject.toml
[tool.poetry]
name = "apicase"
version = "0.1.0"
description = ""
authors = ["guowenhe <18538570410@163.com>"]

[tool.poetry.scripts]
apicase = "apicase.script.cli:main"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
pydantic = "^1.10.4"
jmespath = "^1.0.1"
allure-pytest = "^2.12.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
